shrink deps. cleaning up. integration
