+++ src/templates/dapp/package.json
@@ -8,27 +8,18 @@
 		"@chakra-ui/react": "^2.7.1",
 		"@emotion/react": "^11.11.1",
 		"@emotion/styled": "^11.11.0",
-		"@mui/joy": "^5.0.0-alpha.87",
 		"@pmmmwh/react-refresh-webpack-plugin": "^0.5.3",
 		"@svgr/webpack": "^5.5.0",
-		"@testing-library/jest-dom": "^5.14.1",
-		"@testing-library/react": "^13.0.0",
-		"@testing-library/user-event": "^13.2.1",
 		"@ton-community/blueprint": "^0.11.0",
 		"@tonconnect/protocol": "^2.0.0",
 		"@tonconnect/sdk": "^2.0.5",
-		"@types/jest": "^27.0.1",
 		"@types/node": "^16.7.13",
 		"@types/react": "^18.0.0",
 		"@types/react-dom": "^18.0.0",
 		"antd": "^4.23.6",
 		"babel-jest": "^27.4.2",
-		"babel-loader": "^8.2.3",
-		"babel-plugin-named-asset-import": "^0.3.8",
-		"babel-preset-react-app": "^10.0.1",
+		"babel-loader": "^8.2.3", "babel-preset-react-app": "^10.0.1",
 		"bfj": "^7.0.2",
-		"browserslist": "^4.18.1",
-		"buffer": "^6.0.3",
 		"camelcase": "^6.2.1",
 		"case-sensitive-paths-webpack-plugin": "^2.4.0",
 		"css-loader": "^6.5.1",
@@ -42,9 +33,7 @@
 		"framer-motion": "^10.12.18",
 		"fs-extra": "^10.0.0",
 		"html-webpack-plugin": "^5.5.0",
-		"identity-obj-proxy": "^3.0.0",
 		"jest": "^27.4.3",
-		"jest-resolve": "^27.4.2",
 		"jest-watch-typeahead": "^1.0.0",
 		"mini-css-extract-plugin": "^2.4.5",
 		"postcss": "^8.4.4",
@@ -52,16 +41,12 @@
 		"postcss-loader": "^6.2.1",
 		"postcss-normalize": "^10.0.1",
 		"postcss-preset-env": "^7.0.1",
-		"prompts": "^2.4.2",
 		"react": "^18.2.0",
 		"react-app-polyfill": "^3.0.0",
 		"react-dev-utils": "^12.0.1",
 		"react-dom": "^18.2.0",
-		"react-json-view": "^1.21.3",
 		"react-qr-code": "^2.0.8",
 		"react-refresh": "^0.11.0",
-		"react-router-dom": "^6.14.1",
-		"recast": "^0.23.2",
 		"recoil": "^0.7.6",
 		"resolve": "^1.20.0",
 		"resolve-url-loader": "^4.0.0",
@@ -70,18 +55,16 @@
 		"semver": "^7.3.5",
 		"source-map-loader": "^3.0.0",
 		"style-loader": "^3.3.1",
-		"tailwindcss": "^3.0.2",
 		"terser-webpack-plugin": "^5.2.5",
 		"ton": "^13.3.0",
 		"typescript": "*",
-		"web-vitals": "^2.1.0",
 		"webpack": "^5.64.4",
 		"webpack-dev-server": "^4.6.0",
 		"webpack-manifest-plugin": "^4.0.2",
 		"workbox-webpack-plugin": "^6.4.1"
 	},
 	"devDependencies": {
-		"@trivago/prettier-plugin-sort-imports": "^3.4.0",
+		"@babel/plugin-proposal-private-property-in-object": "^7.21.11",
 		"@typescript-eslint/eslint-plugin": "^5.43.0",
 		"@typescript-eslint/parser": "^5.43.0",
 		"eslint-config-prettier": "^8.5.0",
