+++ src/templates/dapp/src/components/Fields/index.ts
@@ -1,0 +1,9 @@
+export { AddressField } from './Address';
+export { AmountField } from './Amount';
+export { BoolField } from './Bool';
+export { BufferField } from './Buffer';
+export { CellField } from './Cell';
+export { MultiTypeField } from './MultiType';
+export { NullField } from './Null';
+export { StringField } from './String';
+export { UnknownField } from './Unknown';
